tabs look

+++ src/app/views/home/shop/shop-products/shop-products.component.ts
@@ -1,17 +1,18 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { environment } from 'src/environments/environment';
 import { Category, Product, User } from 'src/models';
 import { Company } from 'src/models/company.model';
 import { Interaction, InteractionSearchModel } from 'src/models/interaction.model';
 import { Promotion } from 'src/models/promotion.model';
 import { NavHistoryUX } from 'src/models/UxModel.model';
-import { AccountService, UserService } from 'src/services';
+import { AccountService, OrderService, UserService } from 'src/services';
 import { CompanyService } from 'src/services/company.service';
 import { HomeShopService } from 'src/services/home-shop.service';
 import { InteractionService } from 'src/services/Interaction.service';
 import { ProductService } from 'src/services/product.service';
 import { UxService } from 'src/services/ux.service';
-import { ADMIN, INTERRACTION_TYPE_LIKE, MAX_PAGE_SIZE } from 'src/shared/constants';
+import { ADMIN, DISCOUNT_TYPES, INTERRACTION_TYPE_LIKE, MAX_PAGE_SIZE, ORDER_TYPE_SALES } from 'src/shared/constants';
 
 @Component({
   selector: 'app-shop-products',
@@ -42,6 +43,15 @@ export class ShopProductsComponent implements OnInit {
   nextPage = 999999;
   showShowMore: boolean;
   selectedProduct: Product;
+  heading = `follow a shop.`
+  pendingActionLike: boolean;
+  likeAction: string;
+  showSetUpCompany: boolean;
+  carttItems = 0;
+
+  order: any;
+  fullLink: string;
+  emptyShop: boolean;
   constructor(
     private homeShopService: HomeShopService,
     private productService: ProductService,
@@ -52,6 +62,7 @@ export class ShopProductsComponent implements OnInit {
     private interactionService: InteractionService,
     private accountService: AccountService,
     private userService: UserService,
+    private orderService: OrderService,
 
   ) {
     this.activatedRoute.params.subscribe(r => {
@@ -69,23 +80,60 @@ export class ShopProductsComponent implements OnInit {
   }
 
   ngOnInit() {
+    this.accountService.user.subscribe(data => {
+      this.user = data;
+      // this.getInteractions();
+      if (this.pendingActionLike) {
+        this.onLike(this.likeAction);
+      }
 
-
+    })
+    this.loadCart();
   }
 
-  back(){
+  back() {
     this.router.navigate(['']);
   }
 
+  share() {
+    let nav: any;
+    nav = window.navigator;
+    if (nav.share) {
+      const url = `${environment.BASE_URL}/${this.company.CompanyId}`;
+      nav.share({
+        title: 'Hello there, please checkout.',
+        text: `Hi, please check out *${this.company.Name}*`,
+        url: url,
+      })
+        .then(() => console.log('Successful share'))
+        .catch((error) => console.log('Error sharing', error));
+    } else {
+      this.uxService.showQuickMessage('Product Link Copied to clipboard.');
+
+    }
+  }
+
   getProducts(maxId: number) {
-    this.products = this.productService.getShopProductsState;
-    this.allProducts = this.productService.getShopProductsState;
+    // this.products = this.productService.getShopProductsState;
+    // this.allProducts = this.productService.getShopProductsState;
+    this.uxService.showLoader();
     this.productService.shopProductsObservable.subscribe(data => {
+      this.uxService.hideLoader();
       if (data && data.length) {
         if (JSON.stringify(data) !== JSON.stringify(this.products)) {
           this.products = data;
           this.allProducts = data;
           this.company = this.products[0].Company;
+          this.emptyShop = !this.products[0].ProductId;
+          
+
+          this.fullLink = `${environment.BASE_URL}/${this.company.CompanyId}`;
+          if (this.company && this.company.Followers) {
+            this.company.Followers = this.company.Followers.filter(x => x.SourceName && x.SourceDp)
+          }
+          if (this.company) {
+            this.interactionService.logCompanyPage(this.user, this.company);
+          }
           this.nextPage = this.products[this.products.length - 1]?.Id || 99999;
           this.promotions = this.company.Promotions || [];
           this.promotions.map(x => x.Style = { background: x.Bg, color: x.Color });
@@ -100,6 +148,31 @@ export class ShopProductsComponent implements OnInit {
           this.promotions.map(x => x.Style = { background: x.Bg, color: x.Color });
         }
         this.showShowMore = data.length >= MAX_PAGE_SIZE;
+        if (this.promotions && this.promotions.length) {
+          const promo = this.promotions[0];
+          this.products.forEach(product => {
+            if (promo.PromoType === DISCOUNT_TYPES[0]) {
+              product.SalePrice = (Number(product.RegularPrice) * (Number(promo.DiscountValue) / 100));
+              product.SalePrice = (Number(product.RegularPrice) - (Number(product.SalePrice)));
+              product.Sale = `${promo.DiscountValue} ${promo.DiscountUnits}`
+            }
+            if (promo.PromoType === DISCOUNT_TYPES[1]) {
+              (product.SalePrice = (Number(product.RegularPrice) - (Number(promo.DiscountValue))));
+              product.Sale = `${promo.DiscountValue} ${promo.DiscountUnits}`
+            }
+
+            if (Number(product.SalePrice) < Number(product.RegularPrice)) {
+              product.OnSale = true;
+            }
+          }
+          )
+
+        }
+      } else {
+        // assume the user have no company
+        if (this.user && this.user.UserId && !this.user.CompanyId) {
+          this.showSetUpCompany = true;
+        }
       }
     });
 
@@ -115,6 +188,37 @@ export class ShopProductsComponent implements OnInit {
       });
     }
 
+  }
+
+  cart() {
+    this.router.navigate(['shop/cart']);
+  }
+  loadCart() {
+    this.order = this.orderService.currentOrderValue;
+    if (!this.order) {
+      this.order = {
+        OrdersId: '',
+        OrderNo: 'Shop',
+        CompanyId: this.company.CompanyId,
+        Company: this.company,
+        CustomerId: '',
+        AddressId: '',
+        Notes: '',
+        OrderType: ORDER_TYPE_SALES,
+        Total: 0,
+        Paid: 0,
+        Due: 0,
+        InvoiceDate: new Date(),
+        DueDate: '',
+        CreateUserId: 'shop',
+        ModifyUserId: 'shop',
+        Status: 'Not paid',
+        StatusId: 1,
+        Orderproducts: []
+      }
+      this.orderService.updateOrderState(this.order);
+    }
+    this.carttItems = this.order.Orderproducts && this.order.Orderproducts.length || 0;
   }
 
   loadMore() {
@@ -168,16 +272,15 @@ export class ShopProductsComponent implements OnInit {
   goto(url) {
     this.router.navigate([url]);
   }
+
+
+
+
   onLike(like: string) {
+    this.likeAction = like;
     if (!this.user) {
-      this.uxService.keepNavHistory(
-        {
-          BackToAfterLogin: `/${this.company.Slug || this.company.CompanyId}`,
-          BackTo: this.navHistory && this.navHistory.BackTo || null,
-          ScrollToProduct: null
-        }
-      );
-      this.showAdd = true;
+      this.uxService.openTheQuickLogin();
+      this.pendingActionLike = true;
       return false;
     }
     this.liked = like;
@@ -186,7 +289,7 @@ export class ShopProductsComponent implements OnInit {
         InteractionId: "",
         InteractionType: "Like",
         InteractionSourceId: this.user.UserId,
-        InteractionTargetId: this.company.Slug,
+        InteractionTargetId: this.company.CompanyId,
         TraceId: '1',
         InteractionBody: "Follow",
         Color: '',
@@ -197,8 +300,8 @@ export class ShopProductsComponent implements OnInit {
         InteractionStatus: "Valid",
         ImageUrl: this.company.Dp,
         SourceType: "",
-        SourceName: "",
-        SourceDp: "",
+        SourceName: this.user.Name,
+        SourceDp: this.user.Dp,
         TargetType: "",
         TargetName: "",
         TargetDp: "",
@@ -208,17 +311,43 @@ export class ShopProductsComponent implements OnInit {
       }
 
       this.interactionService.add(this.interaction).subscribe(data => {
-        console.log(data);
+        if (data && data.InteractionId) {
+          this.getInteractions();
+          this.pendingActionLike = false;
+          this.uxService.showQuickMessage('Shop added to favorites.');
+          this.getInteractionSync();
+          this.getProducts(99999999);
+
+        }
       })
     }
 
-    if (like === 'no' && this.interaction.InteractionId && this.interaction.CreateDate) {
+    if (like === 'no' && this.interaction && this.interaction.InteractionId && this.interaction.CreateDate) {
       this.interactionService.delete(this.interaction.InteractionId).subscribe(data => {
-        console.log(data);
-      })
+        this.company.Liked = false;
+        this.uxService.showQuickMessage('Shop removed from favorites.');
+        this.getInteractionSync();
+        this.getInteractions();
+        this.getProducts(99999999);
+      });
     }
 
 
+  }
+
+
+
+  getInteractionSync() {
+    if (!this.user) {
+      return;
+    }
+    const interactionSearchModel: InteractionSearchModel = {
+      InteractionSourceId: this.user.UserId,
+      InteractionTargetId: '',
+      InteractionType: INTERRACTION_TYPE_LIKE,
+      StatusId: 1
+    }
+    this.interactionService.getInteractionsBySourceSync(interactionSearchModel);
   }
 
   getInteractions() {
